requirements.txt
langchain 
langchain-community 
langchain-openai 
langchain-experimental 
neo4j 
wikipedia 
tiktoken 
yfiles_jupyter_graphs